requirements.txt
fastapi
uvicorn
python-multipart
google-genai
google-cloud-storage
pillow
python-dotenv
jinja2
